requirements.txt
torch
torchvision
flask
ultralytics
Pillow